train/object-detection/baseline/frameworks/detectron2/pyproject.toml
[project]
name = "framework-detectron2"
version = "0.1.0"
description = "Standalone Detectron2 runner for the multi-framework orchestrator"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "fvcore>=0.1.5,<0.2.0",
    "opencv-python-headless>=4.9.0",
    "torch @ https://download.pytorch.org/whl/cu121/torch-2.3.0%2Bcu121-cp312-cp312-linux_x86_64.whl",
    "torchvision @ https://download.pytorch.org/whl/cu121/torchvision-0.18.0%2Bcu121-cp312-cp312-linux_x86_64.whl",
    "torchaudio @ https://download.pytorch.org/whl/cu121/torchaudio-2.3.0%2Bcu121-cp312-cp312-linux_x86_64.whl",
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git@v0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["detectron2_runner"]

[tool.uv.extra-build-dependencies]
detectron2 = [
    "torch @ https://download.pytorch.org/whl/cu121/torch-2.3.0%2Bcu121-cp312-cp312-linux_x86_64.whl",
]
